[src/get_ld.c] In the case GMP_NUMB_BITS == 16, cast to unsigned long instead of unsigned int since shift counts can be up to 16 and an int might be on 16 bits only.
[From SVN r13137 (trunk)]

+++ src/get_ld.c
@@ -97,29 +97,29 @@ mpfr_get_ld (mpfr_srcptr x, mpfr_rnd_t rnd_mode)
 #elif GMP_NUMB_BITS == 16
       if (MPFR_LIKELY (denorm == 0))
         {
-          ld.s.manl = tmpmant[0] | ((unsigned int) tmpmant[1] << 16);
-          ld.s.manh = tmpmant[2] | ((unsigned int) tmpmant[3] << 16);
+          ld.s.manl = tmpmant[0] | ((unsigned long) tmpmant[1] << 16);
+          ld.s.manh = tmpmant[2] | ((unsigned long) tmpmant[3] << 16);
         }
       else if (denorm < 16)
         {
           ld.s.manl = (tmpmant[0] >> denorm)
-            | ((unsigned int) tmpmant[1] << (16 - denorm))
-            | ((unsigned int) tmpmant[2] << (32 - denorm));
+            | ((unsigned long) tmpmant[1] << (16 - denorm))
+            | ((unsigned long) tmpmant[2] << (32 - denorm));
           ld.s.manh = (tmpmant[2] >> denorm)
-            | ((unsigned int) tmpmant[3] << (16 - denorm));
+            | ((unsigned long) tmpmant[3] << (16 - denorm));
         }
       else if (denorm < 32)
         {
           ld.s.manl = (tmpmant[1] >> (denorm - 16))
-            | ((unsigned int) tmpmant[2] << (32 - denorm));
+            | ((unsigned long) tmpmant[2] << (32 - denorm));
           if (denorm < 16)
-            ld.s.manl |= ((unsigned int) tmpmant[3] << (48 - denorm));
+            ld.s.manl |= ((unsigned long) tmpmant[3] << (48 - denorm));
           ld.s.manh = tmpmant[3] >> (denorm - 16);
         }
       else if (denorm < 48)
         {
           ld.s.manl = (tmpmant[2] >> (denorm - 32))
-            | ((unsigned int) tmpmant[3] << (64 - denorm));
+            | ((unsigned long) tmpmant[3] << (64 - denorm));
           ld.s.manh = 0;
         }
       else /* 48 <= denorm < 64 */
